release: 0.7.96
