+++ Sources/App/AppSettings.swift
@@ -1,0 +1,193 @@
+import FairApp
+
+@available(macOS 12.0, iOS 15.0, *)
+public struct AppSettingsView: View {
+    public enum Tabs: Hashable {
+        case general, advanced
+    }
+
+    public var body: some View {
+        TabView {
+            GeneralSettingsView()
+                .tabItem {
+                    Label("General", systemImage: "gear")
+                        .symbolVariant(.fill)
+                }
+                .tag(Tabs.general)
+            AdvancedSettingsView()
+                .tabItem {
+                    Label("Advanced", systemImage: "star")
+                        .symbolVariant(.fill)
+                }
+                .tag(Tabs.advanced)
+        }
+        .padding(20)
+        .frame(width: 600)
+    }
+}
+
+@available(macOS 12.0, iOS 15.0, *)
+struct GeneralSettingsView: View {
+    @AppStorage("showPreReleases") private var showPreReleases = false
+//    @AppStorage("controlSize") private var controlSize = 3.0
+    @AppStorage("themeStyle") private var themeStyle = ThemeStyle.system
+    @AppStorage("riskFilter") private var riskFilter = AppRisk.risky
+    @AppStorage("iconBadge") private var iconBadge = true
+    @AppStorage("includeCasks") private var includeCasks = false
+
+    var body: some View {
+        Form {
+            ThemeStylePicker(style: $themeStyle)
+
+//            Slider(value: $controlSize, in: 1...5, step: 1) {
+//                Text("Interface Scale:")
+//            }
+
+            Divider()
+
+            HStack(alignment: .firstTextBaseline) {
+                AppRiskPicker(risk: $riskFilter)
+                riskFilter.riskSummaryText(bold: true)
+                    .textSelection(.enabled)
+                    .font(.body)
+                    .frame(height: 150, alignment: .top)
+            }
+
+            Toggle(isOn: $iconBadge) {
+                Text("Badge App Icon with update count")
+            }
+                .help(Text("Show the number of updates that are available to install."))
+
+            Toggle(isOn: $showPreReleases) {
+                Text("Show Pre-Releases")
+            }
+                .help(Text("Display releases that are not yet production-ready according to the developer's standards."))
+
+//            Text("Pre-releases are experimental versions of software that are less tested than stable versions. They are generally released to garner user feedback and assistance, and so should only be installed by those willing experiment.")
+//                .font(.body)
+//                .multilineTextAlignment(.leading)
+//                .frame(height: 200, alignment: .top)
+
+            #if CASK_SUPPORT
+            Toggle(isOn: $includeCasks) {
+                Text(atx: "Include Casks (requires [`Homebrew`](https://brew.sh/))")
+            }
+            .disabled(includeCasks != true && FileManager.default.isDirectory(url: CaskManager.brewInstallRoot) != true)
+                .help(Text("Adds homebrew Casks to the list of available apps."))
+
+
+            Text("Homebrew Cask is a repository of third-party applications and installers. These packages will be installed using the `brew` command. These packages are not subject to the same sandboxing and security requirements as App Fair fair-ground apps, and so should only be installed from trusted sources.")
+                .font(.body)
+                .multilineTextAlignment(.leading)
+                .frame(height: 200, alignment: .top)
+
+            #endif
+
+
+        }
+        .padding(20)
+    }
+}
+
+
+/// The preferred theme style for the app
+public enum ThemeStyle: String, CaseIterable {
+    case system
+    case light
+    case dark
+}
+
+extension ThemeStyle : Identifiable {
+    public var id: Self { self }
+
+    public var label: Text {
+        switch self {
+        case .system: return Text("System")
+        case .light: return Text("Light")
+        case .dark: return Text("Dark")
+        }
+    }
+
+    public var colorScheme: ColorScheme? {
+        switch self {
+        case .system: return nil
+        case .light: return .light
+        case .dark: return .dark
+        }
+    }
+}
+
+
+@available(macOS 12.0, iOS 15.0, *)
+struct ThemeStylePicker: View {
+    @Binding var style: ThemeStyle
+
+    var body: some View {
+        Picker(selection: $style) {
+            ForEach(ThemeStyle.allCases) { themeStyle in
+                themeStyle.label
+            }
+        } label: {
+            Text("Theme:")
+        }
+        .radioPickerStyle()
+    }
+}
+
+@available(macOS 12.0, iOS 15.0, *)
+struct AppRiskPicker: View {
+    @Binding var risk: AppRisk
+
+    var body: some View {
+        Picker(selection: $risk) {
+            ForEach(AppRisk.allCases) { appRisk in
+                appRisk.riskLabel()
+            }
+        } label: {
+            Text("Risk Exposure:")
+        }
+        .radioPickerStyle()
+    }
+}
+
+@available(macOS 12.0, iOS 15.0, *)
+struct AdvancedSettingsView: View {
+    @EnvironmentObject var fairManager: FairManager
+    @EnvironmentObject var appManager: AppManager
+
+    func checkButton(_ parts: String...) -> some View {
+        EmptyView()
+//        Group {
+//            Image(systemName: "checkmark.square.fill").aspectRatio(contentMode: .fit).foregroundColor(.green)
+//            Image(systemName: "xmark.square.fill").aspectRatio(contentMode: .fit).foregroundColor(.red)
+//        }
+    }
+
+    var body: some View {
+        VStack {
+            Form {
+                HStack {
+                    TextField("Hub", text: fairManager.$hubProvider)
+                    checkButton(fairManager.hubProvider)
+                }
+                HStack {
+                    TextField("Organization", text: fairManager.$hubOrg)
+                    checkButton(fairManager.hubProvider, fairManager.hubOrg)
+                }
+                HStack {
+                    TextField("Repository", text: fairManager.$hubRepo)
+                    checkButton(fairManager.hubProvider, fairManager.hubOrg, fairManager.hubRepo)
+                }
+                HStack {
+                    SecureField("Token", text: fairManager.$hubToken)
+                }
+
+                Text(atx: "The token is optional, and is only needed for development or advanced usage. One can be created at your [GitHub Personal access token](https://github.com/settings/tokens) setting").multilineTextAlignment(.trailing)
+
+                HelpButton(url: "https://github.com/settings/tokens")
+            }
+            .padding(20)
+        }
+    }
+}
+
